@@ -1,24 +1,26 @@
 import React, { useState } from "react";
 import Country from "./Country";
-import { useNavigate } from 'react-router-dom';
+import { useNavigate } from "react-router-dom";
 import ScrollToTop from "./ScrollToTop";
 
 const CountryList = ({ filterCountries, selectedRegion }) => {
-  const [showDetails, setShowDetails] = useState(null)
+  const [showDetails, setShowDetails] = useState(null);
   const history = useNavigate();
 
   const handleToggleDetails = (countryName) => {
     setShowDetails(showDetails === countryName ? null : countryName);
-  }
+  };
 
-  const countriesToRender = filterCountries.filter(country => !selectedRegion || country.region === selectedRegion);
+  const countriesToRender = filterCountries.filter(
+    (country) => !selectedRegion || country.region === selectedRegion
+  );
 
   const navigateToCountryAbout = (country) => {
     history(`/country/${country.name.common}`, { state: { country } });
-  }
+  };
 
   return (
-    <div className='grid grid-container items-baseline gap-16 dark:bg-darkBg bg-white'>
+    <div className="grid grid-container items-baseline gap-16 dark:bg-darkBg bg-white">
       {countriesToRender.map((country, index) => (
         // eslint-disable-next-line
         <a key={index} onClick={() => navigateToCountryAbout(country)}>
@@ -32,7 +34,7 @@ const CountryList = ({ filterCountries, selectedRegion }) => {
       ))}
       <ScrollToTop />
     </div>
-  )
-}
+  );
+};
 
-export default CountryList
+export default CountryList;
